refactor: getCurrentUser 메서드 위치 변경: ProgramService -> MemberService
