+++ src/main/java/com/efub/dhs/domain/member/service/MemberService.java
@@ -1,0 +1,23 @@
+package com.efub.dhs.domain.member.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.efub.dhs.domain.member.entity.Member;
+import com.efub.dhs.domain.member.repository.MemberRepository;
+import com.efub.dhs.global.utils.SecurityUtils;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class MemberService {
+	private final MemberRepository memberRepository;
+
+	public Member getCurrentUser() {
+		String username = SecurityUtils.getCurrentUsername();
+		return memberRepository.findByUsername(username)
+			.orElseThrow(() -> new IllegalArgumentException("해당 아이디의 회원을 찾을 수 없습니다."));
+	}
+}
